+++ OrthogonalMPRViewer.m
@@ -698,7 +698,7 @@ NSString * documentsDirectory();
 	return YES;
 }
 
-- (float)splitView:(NSSplitView *)sender constrainMinCoordinate:(float)proposedMin ofSubviewAt:(int)offset
+- (CGFloat)splitView:(NSSplitView *)sender constrainMinCoordinate:(CGFloat)proposedMin ofSubviewAt:(NSInteger)offset
 {
 	if(offset==0)
 	{
@@ -718,7 +718,7 @@ NSString * documentsDirectory();
 	}
 }
 
-- (float)splitView:(NSSplitView *)sender constrainMaxCoordinate:(float)proposedMax ofSubviewAt:(int)offset
+- (CGFloat)splitView:(NSSplitView *)sender constrainMaxCoordinate:(CGFloat)proposedMax ofSubviewAt:(NSInteger)offset
 {
 	NSRect rect1;
 	rect1 = [sender frame];
